# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GRACE",
        "version": "1.0"
    },
    "paths": {
        "/api/inputs-from-id/{id}": {
            "get": {
                "tags": [
                    "logros"
                ],
                "summary": "Mostrar inputs de determinado título",
                "operationId": "App\\Http\\Controllers\\AchievementsController::all_inputs_from_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar inputs de determinado título."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/achievements/user/{id}": {
            "get": {
                "tags": [
                    "logros"
                ],
                "summary": "Mostrar logros de determinado usuario",
                "operationId": "App\\Http\\Controllers\\AchievementsController::all_achievements_from_user_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar logros de determinado usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "usuarios"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::index",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}